Micrometer support
Additionally, this commit makes sure SmallRye Fault Tolerance doesn't emit
metrics for strategies that a method doesn't define.
