+++ implementation/core/src/main/java/io/smallrye/faulttolerance/core/metrics/CompletionStageMetricsCollector.java
@@ -10,8 +10,9 @@ import io.smallrye.faulttolerance.core.FaultToleranceStrategy;
 import io.smallrye.faulttolerance.core.InvocationContext;
 
 public class CompletionStageMetricsCollector<V> extends MetricsCollector<CompletionStage<V>> {
-    public CompletionStageMetricsCollector(FaultToleranceStrategy<CompletionStage<V>> delegate, MetricsRecorder metrics) {
-        super(delegate, metrics, true);
+    public CompletionStageMetricsCollector(FaultToleranceStrategy<CompletionStage<V>> delegate, MetricsRecorder metrics,
+            boolean hasBulkhead, boolean hasCircuitBreaker, boolean hasRetry, boolean hasTimeout) {
+        super(delegate, metrics, true, hasBulkhead, hasCircuitBreaker, hasRetry, hasTimeout);
     }
 
     @Override
