+++ implementation/fault-tolerance/src/main/java/io/smallrye/faulttolerance/metrics/MetricConstants.java
@@ -1,11 +1,6 @@
 package io.smallrye.faulttolerance.metrics;
 
-import org.eclipse.microprofile.metrics.Metadata;
-import org.eclipse.microprofile.metrics.MetricType;
-import org.eclipse.microprofile.metrics.MetricUnits;
-import org.eclipse.microprofile.metrics.Tag;
-
-public final class MetricConstants {
+final class MetricConstants {
     private MetricConstants() {
         // avoid instantiation
     }
@@ -27,52 +22,4 @@ public final class MetricConstants {
     static final String BULKHEAD_EXECUTIONS_WAITING = "ft.bulkhead.executionsWaiting";
     static final String BULKHEAD_RUNNING_DURATION = "ft.bulkhead.runningDuration";
     static final String BULKHEAD_WAITING_DURATION = "ft.bulkhead.waitingDuration";
-
-    static final Metadata TIMEOUT_EXECUTION_DURATION_METADATA = Metadata.builder()
-            .withName(TIMEOUT_EXECUTION_DURATION)
-            .withType(MetricType.HISTOGRAM)
-            .withUnit(MetricUnits.NANOSECONDS)
-            .build();
-
-    static final Metadata BULKHEAD_RUNNING_DURATION_METADATA = Metadata.builder()
-            .withName(BULKHEAD_RUNNING_DURATION)
-            .withType(MetricType.HISTOGRAM)
-            .withUnit(MetricUnits.NANOSECONDS)
-            .build();
-
-    static final Metadata BULKHEAD_WAITING_DURATION_METADATA = Metadata.builder()
-            .withName(BULKHEAD_WAITING_DURATION)
-            .withType(MetricType.HISTOGRAM)
-            .withUnit(MetricUnits.NANOSECONDS)
-            .build();
-
-    // ---
-
-    static final Tag RESULT_VALUE_RETURNED = new Tag("result", "valueReturned");
-    static final Tag RESULT_EXCEPTION_THROWN = new Tag("result", "exceptionThrown");
-
-    static final Tag FALLBACK_APPLIED = new Tag("fallback", "applied");
-    static final Tag FALLBACK_NOT_APPLIED = new Tag("fallback", "notApplied");
-    static final Tag FALLBACK_NOT_DEFINED = new Tag("fallback", "notDefined");
-
-    static final Tag RETRIED_TRUE = new Tag("retried", "true");
-    static final Tag RETRIED_FALSE = new Tag("retried", "false");
-    static final Tag RETRY_RESULT_VALUE_RETURNED = new Tag("retryResult", "valueReturned");
-    static final Tag RETRY_RESULT_EXCEPTION_NOT_RETRYABLE = new Tag("retryResult", "exceptionNotRetryable");
-    static final Tag RETRY_RESULT_MAX_RETRIES_REACHED = new Tag("retryResult", "maxRetriesReached");
-    static final Tag RETRY_RESULT_MAX_DURATION_REACHED = new Tag("retryResult", "maxDurationReached");
-
-    static final Tag TIMED_OUT_TRUE = new Tag("timedOut", "true");
-    static final Tag TIMED_OUT_FALSE = new Tag("timedOut", "false");
-
-    static final Tag CIRCUIT_BREAKER_RESULT_SUCCESS = new Tag("circuitBreakerResult", "success");
-    static final Tag CIRCUIT_BREAKER_RESULT_FAILURE = new Tag("circuitBreakerResult", "failure");
-    static final Tag CIRCUIT_BREAKER_RESULT_CB_OPEN = new Tag("circuitBreakerResult", "circuitBreakerOpen");
-
-    static final Tag CIRCUIT_BREAKER_STATE_CLOSED = new Tag("state", "closed");
-    static final Tag CIRCUIT_BREAKER_STATE_OPEN = new Tag("state", "open");
-    static final Tag CIRCUIT_BREAKER_STATE_HALF_OPEN = new Tag("state", "halfOpen");
-
-    static final Tag BULKHEAD_RESULT_ACCEPTED = new Tag("bulkheadResult", "accepted");
-    static final Tag BULKHEAD_RESULT_REJECTED = new Tag("bulkheadResult", "rejected");
 }
