+++ pom.xml
@@ -40,6 +40,7 @@
         <version.kotlin>1.7.10</version.kotlin>
         <version.kotlin-dokka>1.7.0</version.kotlin-dokka>
         <version.kotlinx-coroutines>1.6.3</version.kotlinx-coroutines>
+        <version.micrometer-core>1.9.1</version.micrometer-core>
         <version.microprofile-fault-tolerance>3.0</version.microprofile-fault-tolerance>
         <version.microprofile-config-api>2.0</version.microprofile-config-api>
         <version.microprofile-metrics-api>3.0</version.microprofile-metrics-api>
@@ -155,6 +156,11 @@
                 <groupId>com.squareup</groupId>
                 <artifactId>javapoet</artifactId>
                 <version>${version.javapoet}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.micrometer</groupId>
+                <artifactId>micrometer-core</artifactId>
+                <version>${version.micrometer-core}</version>
             </dependency>
             <dependency>
                 <groupId>io.smallrye.common</groupId>
