+++ testsuite/integration/pom.xml
@@ -29,6 +29,10 @@
             <groupId>io.smallrye</groupId>
             <artifactId>smallrye-fault-tolerance-vertx</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.eclipse.microprofile.metrics</groupId>
+            <artifactId>microprofile-metrics-api</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>io.smallrye.config</groupId>
