+++ testsuite/tck/pom.xml
@@ -25,6 +25,10 @@
             <groupId>io.smallrye</groupId>
             <artifactId>smallrye-metrics</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.eclipse.microprofile.metrics</groupId>
+            <artifactId>microprofile-metrics-api</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.eclipse.microprofile.fault-tolerance</groupId>
